Revert "Moved testing of examples and testprojects to tests"
This reverts commit 32876454af06fb563c8d89ba9e1ad285dba2644a.
