+++ .travis.yml
@@ -32,15 +32,16 @@ env:
     # See: http://docs.travis-ci.com/user/encryption-keys/
     - secure: VvwbndU++a2/iNAjk9cd67ATiipDwqcKnxDR4/J2Ik3GH10wHEDUhJ1+MK4WLhedfaOakDOEmarZQS3GwtgvCHO3knpTJuJc8d/bCfZovYuSqdi//BEv4dS7hDt6tQeJfkbBjG0T4yNjPJ3W9R9KDWCy/vj2CUm90BGg2CmxUbg=
   matrix:
-    - CI_FLAGS="-cjlpn 'Various pants self checks'"  # (fkmsr)
-    - CI_FLAGS="-fkmsrcn 'Unit tests for pants and pants-plugins'"  # (jlp)
-    - CI_FLAGS="-fkmsrcjlp 'Python contrib tests'"  # (n)
-    - CI_FLAGS="-fkmsrjlpn -i 6:0 'Python integration tests for pants - shard 1'"  # (c)
-    - CI_FLAGS="-fkmsrjlpn -i 6:1 'Python integration tests for pants - shard 2'"
-    - CI_FLAGS="-fkmsrjlpn -i 6:2 'Python integration tests for pants - shard 3'"
-    - CI_FLAGS="-fkmsrjlpn -i 6:3 'Python integration tests for pants - shard 4'"
-    - CI_FLAGS="-fkmsrjlpn -i 6:4 'Python integration tests for pants - shard 5'"
-    - CI_FLAGS="-fkmsrjlpn -i 6:5 'Python integration tests for pants - shard 6'"
+    - CI_FLAGS="-cjlpnet 'Various pants self checks'"  # (fkmsr)
+    - CI_FLAGS="-fkmsrcjlpn 'Test examples and testprojects'"  # (et)
+    - CI_FLAGS="-fkmsrcnet 'Unit tests for pants and pants-plugins'"  # (jlp)
+    - CI_FLAGS="-fkmsrcjlpet 'Python contrib tests'"  # (n)
+    - CI_FLAGS="-fkmsrjlpnet -i 6:0 'Python integration tests for pants - shard 1'"  # (c)
+    - CI_FLAGS="-fkmsrjlpnet -i 6:1 'Python integration tests for pants - shard 2'"
+    - CI_FLAGS="-fkmsrjlpnet -i 6:2 'Python integration tests for pants - shard 3'"
+    - CI_FLAGS="-fkmsrjlpnet -i 6:3 'Python integration tests for pants - shard 4'"
+    - CI_FLAGS="-fkmsrjlpnet -i 6:4 'Python integration tests for pants - shard 5'"
+    - CI_FLAGS="-fkmsrjlpnet -i 6:5 'Python integration tests for pants - shard 6'"
 
 before_script: |
   ./build-support/bin/ci-sync.sh
