+++ tests/python/pants_test/backend/jvm/tasks/jvm_compile/BUILD
@@ -41,6 +41,14 @@ python_tests(
 )
 
 python_library(
+  name = 'utils',
+  sources = ['utils.py'],
+  dependencies = [
+    '3rdparty/python:pytest',
+  ]
+)
+
+python_library(
   name='base_compile_integration_test',
   sources=['base_compile_integration_test.py'],
   dependencies=[
