+++ tests/python/pants_test/backend/jvm/tasks/jvm_compile/scala/test_scala_compile_integration.py
@@ -9,7 +9,7 @@ import xml.etree.ElementTree as ET
 from contextlib import contextmanager
 
 from pants_test.backend.jvm.tasks.jvm_compile.base_compile_integration_test import BaseCompileIT
-from pants_test.testutils.compile_strategy_utils import provide_compile_strategies
+from pants_test.backend.jvm.tasks.jvm_compile.utils import provide_compile_strategies
 
 
 class ScalaCompileIntegrationTest(BaseCompileIT):
