+++ tests/python/pants_test/tasks/test_jar_publish.py
@@ -22,8 +22,8 @@ from pants.base.exceptions import TaskError
 from pants.scm.scm import Scm
 from pants.util.contextutil import temporary_dir
 from pants.util.dirutil import safe_mkdir, safe_walk
+from pants_test.backend.jvm.tasks.jvm_compile.utils import set_compile_strategies
 from pants_test.tasks.task_test_base import TaskTestBase
-from pants_test.testutils.compile_strategy_utils import set_compile_strategies
 
 
 class JarPublishTest(TaskTestBase):
